## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->

**feat: flox-activations containerize entrypoint**

In the container case we build in some aspects of the activation
context, but some aspects must be determined at runtime since a
user can override the container's `Cmd` at runtime.

When no `Cmd` is provided, we set an "interactive" invocation
type. When arguments are provided either in the containerize
config or at runtime, we supply those as the `ActivateCtx.run_args` with
a "command" invocation type.

This also adds a new field `ActivateCtx.remove_after_reading` to inform
whether the file containing the `ActivateCtx` JSON should be deleted
before the `exec`. In the non-container case, where this is a tempfile,
that makes sense. In the container case, that JSON file is produced by
`mkContainer.nix` and placed in the store, which makes deleting it an
error.

**fix: failing containerize tests**

These tests relied on needing to pass `-c` to a bash shell,
which you no longer need to do (you can just pass commands directly).

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
N/A

<!-- Many thanks! -->

+++ cli/flox-activations/src/cli/activate.rs
@@ -19,6 +19,11 @@ pub struct ActivateArgs {
     /// Path to JSON file containing activation data
     #[arg(long)]
     pub activate_data: PathBuf,
+
+    /// Additional arguments used to provide a command to run.
+    /// NOTE: this is only relevant for containerize activations.
+    #[arg(allow_hyphen_values = true)]
+    pub cmd: Option<Vec<String>>,
 }
 
 pub const FLOX_ENV_LOG_DIR_VAR: &str = "_FLOX_ENV_LOG_DIR";
@@ -33,11 +38,45 @@ pub const FLOX_ACTIVATE_START_SERVICES_VAR: &str = "FLOX_ACTIVATE_START_SERVICES
 impl ActivateArgs {
     pub fn handle(self) -> Result<(), anyhow::Error> {
         let contents = fs::read_to_string(&self.activate_data)?;
-        let data: ActivateCtx = serde_json::from_str(&contents)?;
+        let mut data: ActivateCtx = serde_json::from_str(&contents)?;
 
-        fs::remove_file(&self.activate_data)?;
+        if data.remove_after_reading {
+            fs::remove_file(&self.activate_data)?;
+        }
+
+        // In the case of containerize, you can't bake-in the invocation type or the
+        // `run_args`, so you need to do that detection at runtime. Here we do that
+        // by modifying the `ActivateCtx` passed to us in the container's
+        // EntryPoint.
+        let run_args = self
+            .cmd
+            .as_ref()
+            .or(Some(&data.run_args))
+            .and_then(|args| if args.is_empty() { None } else { Some(args) });
+
+        match (data.invocation_type.as_ref(), run_args) {
+            // This is a container invocation, and we need to set the invocation type
+            // based on the presence of command arguments.
+            (None, None) => data.invocation_type = Some(InvocationType::Interactive),
+            // This is a container invocation, and we need to set the invocation type
+            // based on the presence of command arguments.
+            (None, Some(args)) => {
+                data.invocation_type = Some(InvocationType::Command);
+                data.run_args = args.clone();
+            },
+            // The following two cases are normal shell activations, and don't need
+            // to modify the activation context.
+            (Some(_), None) => {},
+            (Some(_), Some(_)) => {},
+        }
+        // For any case where `invocation_type` is None, we should have detected that above
+        // and set it to Some.
+        let invocation_type = data
+            .invocation_type
+            .expect("invocation type should have been some");
 
         let activate_script_command = Self::assemble_command_for_activate_script(data.clone());
+
         // when output is not a tty, and no command is provided
         // we just print an activation script to stdout
         //
@@ -45,14 +84,14 @@ impl ActivateArgs {
         // e.g. in a .bashrc or .zshrc file:
         //
         //    eval "$(flox activate)"
-        if data.invocation_type == InvocationType::InPlace {
+        if invocation_type == InvocationType::InPlace {
             Self::activate_in_place(activate_script_command, data.shell)?;
 
             return Ok(());
         }
 
         // These functions will only return if exec fails
-        if data.invocation_type == InvocationType::Interactive {
+        if invocation_type == InvocationType::Interactive {
             Self::activate_interactive(activate_script_command)
         } else {
             Self::activate_command(activate_script_command, data.run_args)
@@ -132,7 +171,6 @@ impl ActivateArgs {
     fn assemble_command_for_activate_script(data: ActivateCtx) -> Command {
         let mut exports = HashMap::from([
             (FLOX_ACTIVE_ENVIRONMENTS_VAR, data.flox_active_environments),
-            (FLOX_ENV_LOG_DIR_VAR, data.flox_env_log_dir),
             ("FLOX_PROMPT_COLOR_1", data.prompt_color_1),
             ("FLOX_PROMPT_COLOR_2", data.prompt_color_2),
             // Set `FLOX_PROMPT_ENVIRONMENTS` to the constructed prompt string,
@@ -151,9 +189,13 @@ impl ActivateArgs {
                 FLOX_ACTIVATE_START_SERVICES_VAR,
                 data.flox_activate_start_services.to_string(),
             ),
-            (FLOX_SERVICES_SOCKET_VAR, data.flox_services_socket),
         ]);
-
+        if let Some(log_dir) = data.flox_env_log_dir.as_ref() {
+            exports.insert(FLOX_ENV_LOG_DIR_VAR, log_dir.clone());
+        }
+        if let Some(socket_path) = data.flox_services_socket.as_ref() {
+            exports.insert(FLOX_SERVICES_SOCKET_VAR, socket_path.clone());
+        }
         if let Some(services_to_start) = data.flox_services_to_start {
             exports.insert(FLOX_SERVICES_TO_START_VAR, services_to_start);
         }
@@ -165,9 +207,11 @@ impl ActivateArgs {
         command.envs(exports);
 
         command.arg("--env").arg(&data.env);
-        command
-            .arg("--env-project")
-            .arg(data.env_project.to_string_lossy().to_string());
+        if let Some(env_project) = data.env_project.as_ref() {
+            command
+                .arg("--env-project")
+                .arg(env_project.to_string_lossy().to_string());
+        }
         command
             .arg("--env-cache")
             .arg(data.env_cache.to_string_lossy().to_string());
@@ -176,9 +220,11 @@ impl ActivateArgs {
         // Pass down the activation mode
         command.arg("--mode").arg(data.mode);
 
-        command
-            .arg("--watchdog")
-            .arg(data.watchdog_bin.to_string_lossy().to_string());
+        if let Some(watchdog_bin) = data.watchdog_bin.as_ref() {
+            command
+                .arg("--watchdog")
+                .arg(watchdog_bin.to_string_lossy().to_string());
+        }
 
         command.arg("--shell").arg(data.shell.exe_path());
